+++ tests/cogs/test_admin_cog.py
@@ -227,6 +227,7 @@ class TestCogAdminInfo(InteractionMixin):
             "fields": [
                 {"inline": True, "name": "Format", "value": "Commander"},
                 {"inline": True, "name": "Updated at", "value": ANY},
+                {"inline": False, "name": "Support SpellBot", "value": ANY},
             ],
             "footer": {"text": f"SpellBot Game ID: #SB{game.id}"},
             "thumbnail": {"url": self.settings.THUMB_URL},